Add Device Tree support for x86 and x86_64.
This is a pre-testing commmit.

The major difference between ARM and x86, w.r.t. how the Device Tree Blob (DTB, a.k.a. Flattened Device Tree or FDT) is passed from bootloader to kernel is, that on ARM the pointer is passed in R2, whereas on x86, it's passed in boot_params.setup_header.setup_data.